+++ CHANGELOG.md
@@ -18,7 +18,7 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Minor
 
-- Added Initiate Transfer Confirm ToS Dialog
+- Added Initiate Negotiation Confirm ToS Dialog
 
 #### Patch